// apps/web/src/client/types.gen.ts
export type ItemPublic = {
    title: string;
    description?: (string | null);
    id: string;
    owner_id: string;
    created_at?: (string | null);
};